Changed structure to src folder
