+++ .gitignore
@@ -129,4 +129,4 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-credentials.ini
+config.ini
