+++ config_example.ini
@@ -1,0 +1,10 @@
+# rename this file as 'config.ini' and fill the values
+
+[credentials]
+api_key=your api key
+api_key_secret=your api key secret
+access_token=your access token
+access_token_secret=your access token secret
+
+[url]
+url_pred=http://localhost:3674/predict
